+++ tests/model.rs
@@ -12,3 +12,9 @@ fn test_gl_transmission_format_binary() {
     let format = FileFormat::from_file("fixtures/model/sample.glb").unwrap();
     assert_eq!(format, FileFormat::GlTransmissionFormatBinary);
 }
+
+#[test]
+fn test_google_draco() {
+    let format = FileFormat::from_file("fixtures/model/sample.drc").unwrap();
+    assert_eq!(format, FileFormat::GoogleDraco);
+}